+++ src/Git_Updater/Init.php
@@ -50,6 +50,9 @@ class Init {
 	public function run() {
 		if ( ! static::is_heartbeat() ) {
 			$this->load_hooks();
+
+			new Ignore( 'git-updater-pro', 'git-updater-pro/git-updater-pro.php' );
+			new Ignore( 'git-remote-updater', 'git-remote-updater/git-remote-updater.php' );
 		}
 	}